requirements.txt
absl-py==1.4.0
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.3
# apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
auditwheel==4.0.0
Automat==24.8.1
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bangla==0.0.2
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
Brlapi==0.7.0
Brotli==1.1.0
cached-property==1.5.1
cachetools==5.3.1
certifi==2024.12.14
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.3
clip-anytorch==2.6.0
colorama==0.4.3
colorlog==6.9.0
comm==0.1.3
command-not-found==0.3
constantly==23.10.4
contourpy==1.0.7
coqpit==0.0.17
cramjam==2.9.1
cryptography==45.0.3
cssselect==1.2.0
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.28
dacite==1.8.1
dateparser==1.1.8
dbus-python==1.2.16
dctorch==0.1.2
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
demjson3==3.0.6
dill==0.3.6
distlib==0.3.0
distro==1.9.0
distro-info==0.23+ubuntu1.1
Django==4.2.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
docker==4.1.0
docker-compose==1.25.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
ecdsa==0.19.1
einops==0.8.1
email_validator==2.2.0
entrypoints==0.3
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evidently==0.3.1
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.110.0
fasteners==0.14.1
fastjsonschema==2.18.0
fastparquet==2024.2.0
filelock==3.0.12
Flask==2.3.3
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.39.3
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.2.3
future==0.18.2
g2pkk==0.1.2
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.56.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gunicorn==20.1.0
h11==0.16.0
h5py==3.9.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.29.2
hyperlink==21.0.0
idna==2.8
ImageHash==4.3.1
imageio==2.35.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==24.7.2
inflect==5.6.0
iniconfig==2.1.0
ipykernel==6.25.0
ipython==8.12.2
ipywidgets==8.1.3
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.11
k-diffusion==0.1.1.post1
kaggle==1.6.17
kaleido==0.2.1
keras==2.13.1
keyring==18.0.1
kiwisolver==1.4.4
kornia==0.7.3
kornia_rs==0.1.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.4
libclang==16.0.6
librosa==0.7.2
llvmlite==0.37.0
lockfile==0.12.2
louis==3.12.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.1.3
matplotlib-inline==0.1.6
mecab-python3==1.0.5
mistune==3.0.1
monotonic==1.5
more-itertools==4.2.0
motor==3.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multimethod==1.10
my==1.3.0
mysqlclient==2.2.7
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.7
netifaces==0.10.4
networkx==3.1
nltk==3.9.1
notebook==7.0.0
notebook_shim==0.2.3
num2words==0.5.14
numba==0.54.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
openpyxl==3.1.5
opt-einsum==3.3.0
optional-django==0.3.0
optuna==4.1.0
outcome==1.3.0.post0
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pandas-profiling==3.6.6
pandocfilters==1.5.0
paramiko==2.6.0
parsel==1.9.1
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pbr==6.1.1
pexpect==4.6.0
phik==0.12.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
plotly==5.14.1
pluggy==1.5.0
pooch==1.6.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
propcache==0.2.0
Protego==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
psutil==5.9.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic_core==2.18.2
PyDispatcher==2.0.7
pydyf==0.11.0
pyelftools==0.32
Pygments==2.15.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.3.3
PyMySQL==1.1.1
PyNaCl==1.3.0
pynndescent==0.5.10
pyOpenSSL==25.1.0
pyparsing==3.0.9
pyphen==0.16.0
pypinyin==0.53.0
pyRFC3339==1.1
pyrsistent==0.15.5
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.1
python-apt==2.0.1+ubuntu0.20.4.1
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-slugify==8.0.4
pyttsx3==2.98
pytz==2024.1
PyWavelets==1.4.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.0
qtconsole==5.5.2
QtPy==2.4.1
queuelib==1.7.0
react==4.3.0
redis==5.2.1
referencing==0.30.0
regex==2023.5.5
reportlab==3.5.34
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.4.1
Scrapy==2.11.2
seaborn==0.13.2
SecretStorage==2.3.1
selenium==4.27.1
Send2Trash==1.8.2
sentry-sdk==2.22.0
service-identity==24.2.0
setproctitle==1.3.5
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post4
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.4.1
soxr==0.3.7
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.2
starlette==0.36.3
statsmodels==0.14.0
sympy==1.13.3
systemd-python==234
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
testresources==2.0.2
text-unidecode==1.3
texttable==1.6.2
threadpoolctl==3.1.0
tifffile==2023.7.10
tinycss2==1.2.1
tldextract==5.1.2
tokenizers==0.20.3
tomli==2.0.1
torch==2.4.1
torchaudio==2.4.1
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.19.1
tornado==6.3.2
tqdm==4.65.0
trainer==0.0.20
traitlets==5.9.0
trampoline==0.1.2
transformers==4.46.3
trio==0.27.0
trio-websocket==0.12.2
triton==3.0.0
TTS==0.14.3
Twisted==24.11.0
typeguard==4.4.0
typing_extensions==4.13.2
tzdata==2023.3
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
umap-learn==0.5.1
unattended-upgrades==0.1
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.2.3
usb-creator==0.3.7
uvicorn==0.29.0
uvloop==0.21.0
virtualenv==20.0.17
visions==0.7.6
w3lib==2.2.1
wadllib==1.3.3
wandb==0.19.8
watchfiles==0.24.0
wcwidth==0.2.13
weasyprint==61.2
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
widgetsnbextension==4.0.11
wordcloud==1.9.4
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.1.2
xkit==0.0.0
yarl==1.15.2
ydata-profiling==4.12.0
youtube-dl==2020.3.24
zipp==3.16.2
zope.interface==7.2
zopfli==0.2.3.post1
